[#94] add: Oauth login of github and fixes

+++ trainee/SatyaRajAwasthi/shopping-cart-micro-services/order-service/src/main/java/np/com/esewa/orderservice/resources/OrderResponse.java
@@ -34,8 +34,8 @@ public class OrderResponse {
 
             private String productName;
             private long productId;
-            private long quantity;
-            private long price;
+            private int quantity;
+            private float price;
         }
 
         @Data
